feat: add terminal footer
* add a fixed mono footer statusline with minimal links

* remove forced index pending state so loading only appears when needed
