--- src/components/indexPendingBoundary/index.module.css
@@ -1,9 +1,0 @@
-.loadingPage {
-  flex: 1;
-  width: 100%;
-  min-height: 100dvh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  row-gap: var(--spacing-sm);
-}
